Add files via upload
Classification model:
a. Loan Status Prediction:
i. This is a Property Loan Status Prediction dataset, we have the data of
applicants who had previously applied for the loan based on the property and
applicant-based metrics.
ii. The bank will decide whether to give a loan to the applicant based on some
factors such as Applicant Income, Loan Amount, previous Credit History, Coapplicant Income, etc.
iii. The objective is to build a Machine Learning Model to predict the loan to be
approved or to be rejected for an applicant.
iv. You are free to use any classification model such Logistic Regression, Decision
Tree, RF or XGBoost.
v. Write the performance and key insights on the data in a word doc.